pyproject.toml
[tool.poetry]
name = "expand-sdk"
version = "0.1.0"
description = ""
authors = ["jinyang628 <chenjinyang4192@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
black = "^24.8.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
